Move adapter classes to smaller modules
